requirements.txt
click==8.2.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
requests==2.31.0
python-dotenv==1.0.1
mysql-connector-python==8.3.0
